Chore: Consolidate hooks
