+++ hooks/index.js
@@ -1,0 +1,4 @@
+import useFCL from './useFCL'
+import useLocalStorage from './useLocalStorage'
+
+export { useFCL, useLocalStorage }
